Fix more compiler warnings.

+++ InAppSettingsKit/Models/IASKSettingsStore.h
@@ -22,12 +22,12 @@
 - (void)setBool:(BOOL)value      forKey:(NSString*)key;
 - (void)setFloat:(float)value    forKey:(NSString*)key;
 - (void)setDouble:(double)value  forKey:(NSString*)key;
-- (void)setInteger:(int)value    forKey:(NSString*)key;
+- (void)setInteger:(NSInteger)value    forKey:(NSString*)key;
 - (void)setObject:(id)value      forKey:(NSString*)key;
 - (BOOL)boolForKey:(NSString*)key;
 - (float)floatForKey:(NSString*)key;
 - (double)doubleForKey:(NSString*)key;
-- (int)integerForKey:(NSString*)key;
+- (NSInteger)integerForKey:(NSString*)key;
 - (id)objectForKey:(NSString*)key;
 - (BOOL)synchronize;
 @end
